renamed ROSCameraPublisher.py to ros_camera_publisher.py and ROSLIDARPublisher.py to ros_lidar_publisher.py and adjusted import statement in visualizer operator
