+++ pylot/debug/visualizer_operator.py
@@ -17,8 +17,8 @@ from pylot.planning.world import World
 DEFAULT_VIS_TIME = 30000.0
 
 import rospy
-from .ROSCameraPublisher import ROSCameraPublisher
-from .ROSLIDARPublisher import ROSLIDARPublisher
+from .ros_camera_publisher import ROSCameraPublisher
+from .ros_lidar_publisher import ROSLIDARPublisher
 
 class VisualizerOperator(erdos.Operator):
     """ The `VisualizerOperator` allows developers to see the current state
